Refactored models and removed thickness predictions for stability.
