+++ PINNs/Back_propagation/pytorch_testing/data/MNIST/raw/t10k-images-idx3-ubyte.gz
